python/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.3
aiodecorators==0.2.1
astpretty==3.0.0
asyncstdlib==3.10.5
beautifulsoup4==4.11.1
cbsodata==1.3.4
datapungibea==1.0.2
duckdb==0.9.2
eurostat==1.0.1
fastapi==0.89.1
f90nml==1.4.3
joblib==1.2.0
Jinja2==3.1.3
openpyxl==3.0.10
oracledb==1.3.2
pandas==1.5.2
polars==0.20.13
pyjstat==2.3.0
pymongo==4.3.3
pyyaml==6.0.1
requests-cache==0.9.8
requests==2.28.2
rich==13.0.1
Rx==3.2.0
sdmx1==2.12.1
Sphinx==6.1.3
statsmodels==0.13.5
tenacity==8.1.0
toml==0.10.2
uvicorn==0.20.0
xlrd==2.0.1
xlsx2csv==0.8.2
XlsxWriter==3.0.6
